requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
lxml>=4.9.0
python-dotenv>=1.0.0
pydantic>=2.5.0
sqlalchemy>=2.0.0
pandas>=2.1.0
tqdm>=4.66.0
